Extract admin-only ZoneType
